requirements.txt
signalrcore==0.9.5
requests==2.27.1
python-dotenv==0.19.1
jsonschema==4.0.1